pyproject.toml
[project]
name = "instasaver"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.13"
dependencies = [
    "aiofiles>=24.1.0",
    "aiogram>=3.22.0",
    "aiohttp>=3.12.15",
    "python-dotenv>=1.1.1",
    "yt-dlp>=2025.9.26",
    "gspread>=6.1.4",
    "google-auth>=2.36.0",
]
